@@ -11,37 +11,35 @@
 
     <properties>
         <!-- Compiler and Source settings -->
-        <java.version>11</java.version>
+        <java.version>17</java.version>
         <maven.compiler.target>${java.version}</maven.compiler.target>
         <maven.compiler.source>${java.version}</maven.compiler.source>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 
         <!-- Library versions -->
-        <spring.version>2.4.5</spring.version>
-        <security.version>5.5.0</security.version>
-        <thymeleaf.version>3.0.4.RELEASE</thymeleaf.version>
+        <spring.version>3.0.4</spring.version>
+        <security.version>6.0.2</security.version>
+        <thymeleaf.version>3.1.1.RELEASE</thymeleaf.version>
         <lombok.version>1.18.24</lombok.version>
-        <flyway.version>8.1.0</flyway.version>
-        <databind.version>2.14.0</databind.version>
-        <bouncycastle.version>1.64</bouncycastle.version>
+        <flyway.version>9.16.1</flyway.version>
+        <databind.version>2.14.2</databind.version>
+        <bouncycastle.version>1.70</bouncycastle.version>
         <codec.version>1.15</codec.version>
-        <jQuery.version>3.5.1</jQuery.version>
-        <json.version>20190722</json.version>
-        <commonmark.version>0.16.1</commonmark.version>
-        <font-awesome.version>5.15.2</font-awesome.version>
-        <opencsv.version>5.5</opencsv.version>
-        <jaxb.version>2.3.3</jaxb.version>
-        <junit.version>5.7.0</junit.version>
-        <junitplatform.version>1.7.0</junitplatform.version>
-        <mockito.version>3.6.28</mockito.version>
+        <json.version>20230227</json.version>
+        <commonmark.version>0.21.0</commonmark.version>
+        <opencsv.version>5.7.1</opencsv.version>
+        <jaxb.version>4.0.0</jaxb.version>
+        <junit.version>5.9.2</junit.version>
+        <junitplatform.version>1.9.2</junitplatform.version>
+        <mockito.version>5.2.0</mockito.version>
         <powermock.version>2.0.9</powermock.version>
-        <selenium.version>4.0.0</selenium.version>
-        <javax.validation.version>2.0.1.Final</javax.validation.version>
-        <mysql.version>8.0.24</mysql.version>
-        <h2.version>1.4.200</h2.version>
-        <maven.checkstyle.version>3.1.1</maven.checkstyle.version>
-        <maven.surefire.version>3.0.0-M5</maven.surefire.version>
-        <checkstyle.version>8.37</checkstyle.version>
+        <jakarta.validation.version>3.0.2</jakarta.validation.version>
+        <mysql.version>8.0.32</mysql.version>
+        <h2.version>2.1.214</h2.version>
+        <maven.checkstyle.version>3.2.1</maven.checkstyle.version>
+        <checkstyle.version>9.3</checkstyle.version>
+        <maven.surefire.version>3.0.0</maven.surefire.version>
+        <jacoco.version>0.8.8</jacoco.version>
     </properties>
 
     <dependencies>
