@@ -30,7 +30,7 @@ public interface Sb3ZipRepository extends JpaRepository<Sb3Zip, Integer> {
      * @param experiment The experiment to search for.
      * @return An {@link Optional} zip file.
      */
-    Optional<Sb3Zip> findFirstByUserAndExperiment(User user, Experiment experiment);
+    Optional<Sb3Zip> findFirstByUserAndExperimentOrderByIdDesc(User user, Experiment experiment);
 
     /**
      * Returns a list of ids of all zip files for the given user created during the given experiment, if any exist.
