@@ -17,6 +17,8 @@ import fim.unipassau.de.scratch1984.web.dto.FileDTO;
 import fim.unipassau.de.scratch1984.web.dto.QuestionEventDTO;
 import fim.unipassau.de.scratch1984.web.dto.ResourceEventDTO;
 import fim.unipassau.de.scratch1984.web.dto.Sb3ZipDTO;
+import jakarta.servlet.ServletOutputStream;
+import jakarta.servlet.http.HttpServletResponse;
 import org.json.JSONObject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -26,8 +28,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import javax.servlet.ServletOutputStream;
-import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
@@ -292,7 +292,7 @@ public class EventRestController {
         int experimentId = object.getInt("experiment");
         String secret = object.getString("secret");
 
-        if (!participantService.isInvalidParticipant(userId, experimentId, secret)) {
+        if (!participantService.isInvalidParticipant(userId, experimentId, secret, true)) {
             ids.add(experimentId);
             ids.add(userId);
         }
@@ -311,7 +311,7 @@ public class EventRestController {
     private boolean isInvalidRequest(final String data, final EventDTO eventDTO) {
         JSONObject object = new JSONObject(data);
         String secret = object.getString("secret");
-        return participantService.isInvalidParticipant(eventDTO.getUser(), eventDTO.getExperiment(), secret);
+        return participantService.isInvalidParticipant(eventDTO.getUser(), eventDTO.getExperiment(), secret, true);
     }
 
     /**
