@@ -13,6 +13,7 @@ import fim.unipassau.de.scratch1984.util.NumberParser;
 import fim.unipassau.de.scratch1984.util.PageUtils;
 import fim.unipassau.de.scratch1984.web.dto.ExperimentDTO;
 import fim.unipassau.de.scratch1984.web.dto.UserDTO;
+import jakarta.servlet.http.HttpServletRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -30,7 +31,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.ModelAndView;
 
-import javax.servlet.http.HttpServletRequest;
 import java.util.ResourceBundle;
 
 /**
@@ -393,7 +393,7 @@ public class HomeController {
             LOGGER.error("Cannot finish experiment with secret null or blank!");
             return true;
         } else {
-            return participantService.isInvalidParticipant(userId, experimentId, secret);
+            return participantService.isInvalidParticipant(userId, experimentId, secret, false);
         }
     }
 
