@@ -17,15 +17,14 @@ import fim.unipassau.de.scratch1984.persistence.repository.UserRepository;
 import fim.unipassau.de.scratch1984.util.enums.Language;
 import fim.unipassau.de.scratch1984.util.enums.Role;
 import fim.unipassau.de.scratch1984.web.dto.ParticipantDTO;
+import jakarta.persistence.EntityNotFoundException;
+import jakarta.validation.ConstraintViolationException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-
-import javax.persistence.EntityNotFoundException;
-import javax.validation.ConstraintViolationException;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
@@ -108,8 +107,8 @@ public class ParticipantServiceTest {
 
     @Test
     public void testGetParticipant() {
-        when(userRepository.getOne(ID)).thenReturn(user);
-        when(experimentRepository.getOne(ID)).thenReturn(experiment1);
+        when(userRepository.getReferenceById(ID)).thenReturn(user);
+        when(experimentRepository.getReferenceById(ID)).thenReturn(experiment1);
         when(participantRepository.findByUserAndExperiment(user, experiment1)).thenReturn(Optional.of(participant1));
         ParticipantDTO participantDTO = participantService.getParticipant(ID, ID);
         assertAll(
@@ -118,33 +117,33 @@ public class ParticipantServiceTest {
                 () -> assertNull(participantDTO.getStart()),
                 () -> assertNull(participantDTO.getEnd())
         );
-        verify(userRepository).getOne(ID);
-        verify(experimentRepository).getOne(ID);
+        verify(userRepository).getReferenceById(ID);
+        verify(experimentRepository).getReferenceById(ID);
         verify(participantRepository).findByUserAndExperiment(user, experiment1);
     }
 
     @Test
     public void testGetParticipantNull() {
-        when(userRepository.getOne(ID)).thenReturn(user);
-        when(experimentRepository.getOne(ID)).thenReturn(experiment1);
+        when(userRepository.getReferenceById(ID)).thenReturn(user);
+        when(experimentRepository.getReferenceById(ID)).thenReturn(experiment1);
         assertThrows(NotFoundException.class,
                 () -> participantService.getParticipant(ID, ID)
         );
-        verify(userRepository).getOne(ID);
-        verify(experimentRepository).getOne(ID);
+        verify(userRepository).getReferenceById(ID);
+        verify(experimentRepository).getReferenceById(ID);
         verify(participantRepository).findByUserAndExperiment(user, experiment1);
     }
 
     @Test
     public void testGetParticipantEntityNotFound() {
-        when(userRepository.getOne(ID)).thenReturn(user);
-        when(experimentRepository.getOne(ID)).thenReturn(experiment1);
+        when(userRepository.getReferenceById(ID)).thenReturn(user);
+        when(experimentRepository.getReferenceById(ID)).thenReturn(experiment1);
         when(participantRepository.findByUserAndExperiment(user, experiment1)).thenThrow(EntityNotFoundException.class);
         assertThrows(NotFoundException.class,
                 () -> participantService.getParticipant(ID, ID)
         );
-        verify(userRepository).getOne(ID);
-        verify(experimentRepository).getOne(ID);
+        verify(userRepository).getReferenceById(ID);
+        verify(experimentRepository).getReferenceById(ID);
         verify(participantRepository).findByUserAndExperiment(user, experiment1);
     }
 
@@ -153,8 +152,8 @@ public class ParticipantServiceTest {
         assertThrows(IllegalArgumentException.class,
                 () -> participantService.getParticipant(ID, 0)
         );
-        verify(userRepository, never()).getOne(anyInt());
-        verify(experimentRepository, never()).getOne(anyInt());
+        verify(userRepository, never()).getReferenceById(anyInt());
+        verify(experimentRepository, never()).getReferenceById(anyInt());
         verify(participantRepository, never()).findByUserAndExperiment(any(), any());
     }
 
@@ -163,20 +162,20 @@ public class ParticipantServiceTest {
         assertThrows(IllegalArgumentException.class,
                 () -> participantService.getParticipant(-1, ID)
         );
-        verify(userRepository, never()).getOne(anyInt());
-        verify(experimentRepository, never()).getOne(anyInt());
+        verify(userRepository, never()).getReferenceById(anyInt());
+        verify(experimentRepository, never()).getReferenceById(anyInt());
         verify(participantRepository, never()).findByUserAndExperiment(any(), any());
     }
 
     @Test
     public void testSaveParticipants() {
-        when(courseRepository.getOne(ID)).thenReturn(course);
-        when(experimentRepository.getOne(ID)).thenReturn(experiment2);
+        when(courseRepository.getReferenceById(ID)).thenReturn(course);
+        when(experimentRepository.getReferenceById(ID)).thenReturn(experiment2);
         when(courseExperimentRepository.existsByCourseAndExperiment(course, experiment2)).thenReturn(true);
         when(courseParticipantRepository.findAllByCourse(course)).thenReturn(courseParticipants);
         assertDoesNotThrow(() -> participantService.saveParticipants(ID, ID));
-        verify(courseRepository).getOne(ID);
-        verify(experimentRepository).getOne(ID);
+        verify(courseRepository).getReferenceById(ID);
+        verify(experimentRepository).getReferenceById(ID);
         verify(courseExperimentRepository).existsByCourseAndExperiment(course, experiment2);
         verify(courseParticipantRepository).findAllByCourse(course);
         verify(participantRepository, times(3)).save(any());
@@ -185,14 +184,14 @@ public class ParticipantServiceTest {
 
     @Test
     public void testSaveParticipantsConstraintViolation() {
-        when(courseRepository.getOne(ID)).thenReturn(course);
-        when(experimentRepository.getOne(ID)).thenReturn(experiment2);
+        when(courseRepository.getReferenceById(ID)).thenReturn(course);
+        when(experimentRepository.getReferenceById(ID)).thenReturn(experiment2);
         when(courseExperimentRepository.existsByCourseAndExperiment(course, experiment2)).thenReturn(true);
         when(courseParticipantRepository.findAllByCourse(course)).thenReturn(courseParticipants);
         when(participantRepository.save(any())).thenThrow(ConstraintViolationException.class);
         assertThrows(StoreException.class, () -> participantService.saveParticipants(ID, ID));
-        verify(courseRepository).getOne(ID);
-        verify(experimentRepository).getOne(ID);
+        verify(courseRepository).getReferenceById(ID);
+        verify(experimentRepository).getReferenceById(ID);
         verify(courseExperimentRepository).existsByCourseAndExperiment(course, experiment2);
         verify(courseParticipantRepository).findAllByCourse(course);
         verify(participantRepository).save(any());
@@ -201,13 +200,13 @@ public class ParticipantServiceTest {
 
     @Test
     public void testSaveParticipantsEntityNotFound() {
-        when(courseRepository.getOne(ID)).thenReturn(course);
-        when(experimentRepository.getOne(ID)).thenReturn(experiment2);
+        when(courseRepository.getReferenceById(ID)).thenReturn(course);
+        when(experimentRepository.getReferenceById(ID)).thenReturn(experiment2);
         when(courseExperimentRepository.existsByCourseAndExperiment(course,
                 experiment2)).thenThrow(EntityNotFoundException.class);
         assertThrows(NotFoundException.class, () -> participantService.saveParticipants(ID, ID));
-        verify(courseRepository).getOne(ID);
-        verify(experimentRepository).getOne(ID);
+        verify(courseRepository).getReferenceById(ID);
+        verify(experimentRepository).getReferenceById(ID);
         verify(courseExperimentRepository).existsByCourseAndExperiment(course, experiment2);
         verify(courseParticipantRepository, never()).findAllByCourse(any());
         verify(participantRepository, never()).save(any());
@@ -216,11 +215,11 @@ public class ParticipantServiceTest {
 
     @Test
     public void testSaveParticipantsExperimentNotCourseExperiment() {
-        when(courseRepository.getOne(ID)).thenReturn(course);
-        when(experimentRepository.getOne(ID)).thenReturn(experiment2);
+        when(courseRepository.getReferenceById(ID)).thenReturn(course);
+        when(experimentRepository.getReferenceById(ID)).thenReturn(experiment2);
         assertThrows(IllegalStateException.class, () -> participantService.saveParticipants(ID, ID));
-        verify(courseRepository).getOne(ID);
-        verify(experimentRepository).getOne(ID);
+        verify(courseRepository).getReferenceById(ID);
+        verify(experimentRepository).getReferenceById(ID);
         verify(courseParticipantRepository, never()).findAllByCourse(any());
         verify(participantRepository, never()).save(any());
         verify(userRepository, never()).save(any());
@@ -229,11 +228,11 @@ public class ParticipantServiceTest {
     @Test
     public void testSaveParticipantsExperimentInactive() {
         experiment1.setActive(false);
-        when(courseRepository.getOne(ID)).thenReturn(course);
-        when(experimentRepository.getOne(ID)).thenReturn(experiment1);
+        when(courseRepository.getReferenceById(ID)).thenReturn(course);
+        when(experimentRepository.getReferenceById(ID)).thenReturn(experiment1);
         assertThrows(IllegalStateException.class, () -> participantService.saveParticipants(ID, ID));
-        verify(courseRepository).getOne(ID);
-        verify(experimentRepository).getOne(ID);
+        verify(courseRepository).getReferenceById(ID);
+        verify(experimentRepository).getReferenceById(ID);
         verify(courseParticipantRepository, never()).findAllByCourse(any());
         verify(participantRepository, never()).save(any());
         verify(userRepository, never()).save(any());
@@ -242,11 +241,11 @@ public class ParticipantServiceTest {
     @Test
     public void testSaveParticipantsCourseInactive() {
         course.setActive(false);
-        when(courseRepository.getOne(ID)).thenReturn(course);
-        when(experimentRepository.getOne(ID)).thenReturn(experiment1);
+        when(courseRepository.getReferenceById(ID)).thenReturn(course);
+        when(experimentRepository.getReferenceById(ID)).thenReturn(experiment1);
         assertThrows(IllegalStateException.class, () -> participantService.saveParticipants(ID, ID));
-        verify(courseRepository).getOne(ID);
-        verify(experimentRepository).getOne(ID);
+        verify(courseRepository).getReferenceById(ID);
+        verify(experimentRepository).getReferenceById(ID);
         verify(courseParticipantRepository, never()).findAllByCourse(any());
         verify(participantRepository, never()).save(any());
         verify(userRepository, never()).save(any());
@@ -258,8 +257,8 @@ public class ParticipantServiceTest {
                 () -> assertThrows(IllegalArgumentException.class, () -> participantService.saveParticipants(0, ID)),
                 () -> assertThrows(IllegalArgumentException.class, () -> participantService.saveParticipants(ID, -1))
         );
-        verify(courseRepository, never()).getOne(anyInt());
-        verify(experimentRepository, never()).getOne(anyInt());
+        verify(courseRepository, never()).getReferenceById(anyInt());
+        verify(experimentRepository, never()).getReferenceById(anyInt());
         verify(courseParticipantRepository, never()).findAllByCourse(any());
         verify(participantRepository, never()).save(any());
         verify(userRepository, never()).save(any());
@@ -267,11 +266,11 @@ public class ParticipantServiceTest {
 
     @Test
     public void testSaveParticipant() {
-        when(userRepository.getOne(ID)).thenReturn(user);
-        when(experimentRepository.getOne(ID)).thenReturn(experiment1);
+        when(userRepository.getReferenceById(ID)).thenReturn(user);
+        when(experimentRepository.getReferenceById(ID)).thenReturn(experiment1);
         assertDoesNotThrow(() -> participantService.saveParticipant(ID, ID));
-        verify(userRepository).getOne(ID);
-        verify(experimentRepository).getOne(ID);
+        verify(userRepository).getReferenceById(ID);
+        verify(experimentRepository).getReferenceById(ID);
         verify(participantRepository).save(any());
     }
 
@@ -281,8 +280,8 @@ public class ParticipantServiceTest {
         assertThrows(NotFoundException.class,
                 () -> participantService.saveParticipant(ID, ID)
         );
-        verify(userRepository).getOne(ID);
-        verify(experimentRepository).getOne(ID);
+        verify(userRepository).getReferenceById(ID);
+        verify(experimentRepository).getReferenceById(ID);
         verify(participantRepository).save(any());
     }
 
@@ -292,8 +291,8 @@ public class ParticipantServiceTest {
         assertThrows(StoreException.class,
                 () -> participantService.saveParticipant(ID, ID)
         );
-        verify(userRepository).getOne(ID);
-        verify(experimentRepository).getOne(ID);
+        verify(userRepository).getReferenceById(ID);
+        verify(experimentRepository).getReferenceById(ID);
         verify(participantRepository).save(any());
     }
 
@@ -303,40 +302,40 @@ public class ParticipantServiceTest {
                 () -> assertThrows(IllegalArgumentException.class, () -> participantService.saveParticipant(ID, 0)),
                 () -> assertThrows(IllegalArgumentException.class, () -> participantService.saveParticipant(-1, ID))
         );
-        verify(userRepository, never()).getOne(anyInt());
-        verify(experimentRepository, never()).getOne(anyInt());
+        verify(userRepository, never()).getReferenceById(anyInt());
+        verify(experimentRepository, never()).getReferenceById(anyInt());
         verify(participantRepository, never()).save(any());
     }
 
     @Test
     public void testUpdateParticipant() {
-        when(userRepository.getOne(ID)).thenReturn(user);
-        when(experimentRepository.getOne(ID)).thenReturn(experiment1);
+        when(userRepository.getReferenceById(ID)).thenReturn(user);
+        when(experimentRepository.getReferenceById(ID)).thenReturn(experiment1);
         assertTrue(participantService.updateParticipant(participantDTO));
-        verify(userRepository).getOne(ID);
-        verify(experimentRepository).getOne(ID);
+        verify(userRepository).getReferenceById(ID);
+        verify(experimentRepository).getReferenceById(ID);
         verify(participantRepository).save(any());
     }
 
     @Test
     public void testUpdateParticipantEntityNotFound() {
-        when(userRepository.getOne(ID)).thenReturn(user);
-        when(experimentRepository.getOne(ID)).thenReturn(experiment1);
+        when(userRepository.getReferenceById(ID)).thenReturn(user);
+        when(experimentRepository.getReferenceById(ID)).thenReturn(experiment1);
         when(participantRepository.save(any())).thenThrow(EntityNotFoundException.class);
         assertFalse(participantService.updateParticipant(participantDTO));
-        verify(userRepository).getOne(ID);
-        verify(experimentRepository).getOne(ID);
+        verify(userRepository).getReferenceById(ID);
+        verify(experimentRepository).getReferenceById(ID);
         verify(participantRepository).save(any());
     }
 
     @Test
     public void testUpdateParticipantConstraintViolation() {
-        when(userRepository.getOne(ID)).thenReturn(user);
-        when(experimentRepository.getOne(ID)).thenReturn(experiment1);
+        when(userRepository.getReferenceById(ID)).thenReturn(user);
+        when(experimentRepository.getReferenceById(ID)).thenReturn(experiment1);
         when(participantRepository.save(any())).thenThrow(ConstraintViolationException.class);
         assertFalse(participantService.updateParticipant(participantDTO));
-        verify(userRepository).getOne(ID);
-        verify(experimentRepository).getOne(ID);
+        verify(userRepository).getReferenceById(ID);
+        verify(experimentRepository).getReferenceById(ID);
         verify(participantRepository).save(any());
     }
 
@@ -346,8 +345,8 @@ public class ParticipantServiceTest {
         assertThrows(IllegalArgumentException.class,
                 () -> participantService.updateParticipant(participantDTO)
         );
-        verify(userRepository, never()).getOne(anyInt());
-        verify(experimentRepository, never()).getOne(anyInt());
+        verify(userRepository, never()).getReferenceById(anyInt());
+        verify(experimentRepository, never()).getReferenceById(anyInt());
         verify(participantRepository, never()).save(any());
     }
 
@@ -357,18 +356,18 @@ public class ParticipantServiceTest {
         assertThrows(IllegalArgumentException.class,
                 () -> participantService.updateParticipant(participantDTO)
         );
-        verify(userRepository, never()).getOne(anyInt());
-        verify(experimentRepository, never()).getOne(anyInt());
+        verify(userRepository, never()).getReferenceById(anyInt());
+        verify(experimentRepository, never()).getReferenceById(anyInt());
         verify(participantRepository, never()).save(any());
     }
 
     @Test
     public void testDeactivateParticipantAccounts() {
-        when(experimentRepository.getOne(ID)).thenReturn(experiment1);
+        when(experimentRepository.getReferenceById(ID)).thenReturn(experiment1);
         when(participantRepository.findAllByExperiment(experiment1)).thenReturn(participantList);
         when(userRepository.findById(any(Integer.class))).thenReturn(java.util.Optional.of(user));
         assertDoesNotThrow(() -> participantService.deactivateParticipantAccounts(ID));
-        verify(experimentRepository).getOne(ID);
+        verify(experimentRepository).getReferenceById(ID);
         verify(participantRepository).findAllByExperiment(experiment1);
         verify(userRepository, times(5)).findById(any(Integer.class));
         verify(userRepository, times(5)).save(user);
@@ -376,12 +375,12 @@ public class ParticipantServiceTest {
 
     @Test
     public void testDeactivateParticipantAccountsSimultaneousParticipation() {
-        when(experimentRepository.getOne(ID)).thenReturn(experiment1);
+        when(experimentRepository.getReferenceById(ID)).thenReturn(experiment1);
         when(participantRepository.findAllByExperiment(experiment1)).thenReturn(participantList);
         when(userRepository.findById(any(Integer.class))).thenReturn(java.util.Optional.of(user));
         when(participantRepository.findAllByEndIsNullAndUser(any())).thenReturn(participantList);
         assertDoesNotThrow(() -> participantService.deactivateParticipantAccounts(ID));
-        verify(experimentRepository).getOne(ID);
+        verify(experimentRepository).getReferenceById(ID);
         verify(participantRepository).findAllByExperiment(experiment1);
         verify(userRepository, times(5)).findById(any(Integer.class));
         verify(userRepository, never()).save(any());
@@ -389,12 +388,12 @@ public class ParticipantServiceTest {
 
     @Test
     public void testDeactivateParticipantAccountsIllegalState() {
-        when(experimentRepository.getOne(ID)).thenReturn(experiment1);
+        when(experimentRepository.getReferenceById(ID)).thenReturn(experiment1);
         when(participantRepository.findAllByExperiment(experiment1)).thenReturn(participantList);
         assertThrows(IllegalStateException.class,
                 () -> participantService.deactivateParticipantAccounts(ID)
         );
-        verify(experimentRepository).getOne(ID);
+        verify(experimentRepository).getReferenceById(ID);
         verify(participantRepository).findAllByExperiment(experiment1);
         verify(userRepository).findById(any(Integer.class));
         verify(userRepository, never()).save(user);
@@ -402,12 +401,12 @@ public class ParticipantServiceTest {
 
     @Test
     public void testDeactivateParticipantAccountsNotFound() {
-        when(experimentRepository.getOne(ID)).thenReturn(experiment1);
+        when(experimentRepository.getReferenceById(ID)).thenReturn(experiment1);
         when(participantRepository.findAllByExperiment(experiment1)).thenThrow(EntityNotFoundException.class);
         assertThrows(NotFoundException.class,
                 () -> participantService.deactivateParticipantAccounts(ID)
         );
-        verify(experimentRepository).getOne(ID);
+        verify(experimentRepository).getReferenceById(ID);
         verify(participantRepository).findAllByExperiment(experiment1);
         verify(userRepository, never()).findById(any(Integer.class));
         verify(userRepository, never()).save(user);
@@ -418,7 +417,7 @@ public class ParticipantServiceTest {
         assertThrows(IllegalArgumentException.class,
                 () -> participantService.deactivateParticipantAccounts(0)
         );
-        verify(experimentRepository, never()).getOne(ID);
+        verify(experimentRepository, never()).getReferenceById(ID);
         verify(participantRepository, never()).findAllByExperiment(experiment1);
         verify(userRepository, never()).findById(any(Integer.class));
         verify(userRepository, never()).save(user);
@@ -426,7 +425,7 @@ public class ParticipantServiceTest {
 
     @Test
     public void testGetExperimentIdsForParticipant() {
-        when(userRepository.getOne(ID)).thenReturn(user);
+        when(userRepository.getReferenceById(ID)).thenReturn(user);
         when(participantRepository.findAllByUser(user)).thenReturn(participantList);
         HashMap<Integer, String> experiments = participantService.getExperimentInfoForParticipant(ID);
         assertAll(
@@ -442,18 +441,18 @@ public class ParticipantServiceTest {
                 () -> assertTrue(experiments.containsKey(5)),
                 () -> assertTrue(experiments.containsValue("Title 4"))
         );
-        verify(userRepository).getOne(ID);
+        verify(userRepository).getReferenceById(ID);
         verify(participantRepository).findAllByUser(user);
     }
 
     @Test
     public void testGetExperimentIdsForParticipantNotFound() {
-        when(userRepository.getOne(ID)).thenReturn(user);
+        when(userRepository.getReferenceById(ID)).thenReturn(user);
         when(participantRepository.findAllByUser(user)).thenThrow(EntityNotFoundException.class);
         assertThrows(NotFoundException.class,
                 () -> participantService.getExperimentInfoForParticipant(ID)
         );
-        verify(userRepository).getOne(ID);
+        verify(userRepository).getReferenceById(ID);
         verify(participantRepository).findAllByUser(user);
     }
 
@@ -462,7 +461,7 @@ public class ParticipantServiceTest {
         assertThrows(IllegalArgumentException.class,
                 () -> participantService.getExperimentInfoForParticipant(0)
         );
-        verify(userRepository, never()).getOne(ID);
+        verify(userRepository, never()).getReferenceById(ID);
         verify(participantRepository, never()).findAllByUser(user);
     }
 
@@ -490,10 +489,10 @@ public class ParticipantServiceTest {
 
     @Test
     public void testSimultaneousParticipation() {
-        when(userRepository.getOne(ID)).thenReturn(user);
+        when(userRepository.getReferenceById(ID)).thenReturn(user);
         when(participantRepository.findAllByEndIsNullAndUser(user)).thenReturn(participantList);
         assertTrue(participantService.simultaneousParticipation(ID));
-        verify(userRepository).getOne(ID);
+        verify(userRepository).getReferenceById(ID);
         verify(participantRepository).findAllByEndIsNullAndUser(user);
     }
 
@@ -501,21 +500,21 @@ public class ParticipantServiceTest {
     public void testSimultaneousParticipationFalse() {
         List<Participant> oneParticipation = new ArrayList<>();
         oneParticipation.add(new Participant());
-        when(userRepository.getOne(ID)).thenReturn(user);
+        when(userRepository.getReferenceById(ID)).thenReturn(user);
         when(participantRepository.findAllByEndIsNullAndUser(user)).thenReturn(oneParticipation);
         assertFalse(participantService.simultaneousParticipation(ID));
-        verify(userRepository).getOne(ID);
+        verify(userRepository).getReferenceById(ID);
         verify(participantRepository).findAllByEndIsNullAndUser(user);
     }
 
     @Test
     public void testSimultaneousParticipationNotFound() {
-        when(userRepository.getOne(ID)).thenReturn(user);
+        when(userRepository.getReferenceById(ID)).thenReturn(user);
         when(participantRepository.findAllByEndIsNullAndUser(user)).thenThrow(EntityNotFoundException.class);
         assertThrows(NotFoundException.class,
                 () -> participantService.simultaneousParticipation(ID)
         );
-        verify(userRepository).getOne(ID);
+        verify(userRepository).getReferenceById(ID);
         verify(participantRepository).findAllByEndIsNullAndUser(user);
     }
 
@@ -524,42 +523,53 @@ public class ParticipantServiceTest {
         assertThrows(IllegalArgumentException.class,
                 () -> participantService.simultaneousParticipation(0)
         );
-        verify(userRepository, never()).getOne(anyInt());
+        verify(userRepository, never()).getReferenceById(anyInt());
         verify(participantRepository, never()).findAllByEndIsNullAndUser(any());
     }
 
     @Test
     public void testIsInvalidParticipant() {
         user.setActive(true);
-        when(userRepository.getOne(ID)).thenReturn(user);
-        when(experimentRepository.getOne(ID)).thenReturn(experiment1);
+        when(userRepository.getReferenceById(ID)).thenReturn(user);
+        when(experimentRepository.getReferenceById(ID)).thenReturn(experiment1);
         when(participantRepository.findByUserAndExperiment(user, experiment1)).thenReturn(Optional.of(participant1));
-        assertFalse(participantService.isInvalidParticipant(ID, ID, SECRET));
-        verify(userRepository).getOne(ID);
-        verify(experimentRepository).getOne(ID);
+        assertFalse(participantService.isInvalidParticipant(ID, ID, SECRET, true));
+        verify(userRepository).getReferenceById(ID);
+        verify(experimentRepository).getReferenceById(ID);
+        verify(participantRepository).findByUserAndExperiment(user, experiment1);
+    }
+
+    @Test
+    public void testIsInvalidParticipantInactive() {
+        when(userRepository.getReferenceById(ID)).thenReturn(user);
+        when(experimentRepository.getReferenceById(ID)).thenReturn(experiment1);
+        when(participantRepository.findByUserAndExperiment(user, experiment1)).thenReturn(Optional.of(participant1));
+        assertFalse(participantService.isInvalidParticipant(ID, ID, SECRET, false));
+        verify(userRepository).getReferenceById(ID);
+        verify(experimentRepository).getReferenceById(ID);
         verify(participantRepository).findByUserAndExperiment(user, experiment1);
     }
 
     @Test
     public void testIsInvalidParticipantSecretNotMatching() {
         user.setActive(true);
-        when(userRepository.getOne(ID)).thenReturn(user);
-        when(experimentRepository.getOne(ID)).thenReturn(experiment1);
+        when(userRepository.getReferenceById(ID)).thenReturn(user);
+        when(experimentRepository.getReferenceById(ID)).thenReturn(experiment1);
         when(participantRepository.findByUserAndExperiment(user, experiment1)).thenReturn(Optional.of(participant1));
-        assertTrue(participantService.isInvalidParticipant(ID, ID, PASSWORD));
-        verify(userRepository).getOne(ID);
-        verify(experimentRepository).getOne(ID);
+        assertTrue(participantService.isInvalidParticipant(ID, ID, PASSWORD, true));
+        verify(userRepository).getReferenceById(ID);
+        verify(experimentRepository).getReferenceById(ID);
         verify(participantRepository).findByUserAndExperiment(user, experiment1);
     }
 
     @Test
     public void testIsInvalidParticipantUserInactive() {
-        when(userRepository.getOne(ID)).thenReturn(user);
-        when(experimentRepository.getOne(ID)).thenReturn(experiment1);
+        when(userRepository.getReferenceById(ID)).thenReturn(user);
+        when(experimentRepository.getReferenceById(ID)).thenReturn(experiment1);
         when(participantRepository.findByUserAndExperiment(user, experiment1)).thenReturn(Optional.of(participant1));
-        assertTrue(participantService.isInvalidParticipant(ID, ID, SECRET));
-        verify(userRepository).getOne(ID);
-        verify(experimentRepository).getOne(ID);
+        assertTrue(participantService.isInvalidParticipant(ID, ID, SECRET, true));
+        verify(userRepository).getReferenceById(ID);
+        verify(experimentRepository).getReferenceById(ID);
         verify(participantRepository).findByUserAndExperiment(user, experiment1);
     }
 
@@ -567,73 +577,73 @@ public class ParticipantServiceTest {
     public void testIsInvalidParticipantExperimentInactive() {
         user.setActive(true);
         experiment1.setActive(false);
-        when(userRepository.getOne(ID)).thenReturn(user);
-        when(experimentRepository.getOne(ID)).thenReturn(experiment1);
+        when(userRepository.getReferenceById(ID)).thenReturn(user);
+        when(experimentRepository.getReferenceById(ID)).thenReturn(experiment1);
         when(participantRepository.findByUserAndExperiment(user, experiment1)).thenReturn(Optional.of(participant1));
-        assertTrue(participantService.isInvalidParticipant(ID, ID, SECRET));
-        verify(userRepository).getOne(ID);
-        verify(experimentRepository).getOne(ID);
+        assertTrue(participantService.isInvalidParticipant(ID, ID, SECRET, true));
+        verify(userRepository).getReferenceById(ID);
+        verify(experimentRepository).getReferenceById(ID);
         verify(participantRepository).findByUserAndExperiment(user, experiment1);
     }
 
     @Test
     public void testIsInvalidParticipantNoParticipant() {
-        when(userRepository.getOne(ID)).thenReturn(user);
-        when(experimentRepository.getOne(ID)).thenReturn(experiment1);
-        assertTrue(participantService.isInvalidParticipant(ID, ID, SECRET));
-        verify(userRepository).getOne(ID);
-        verify(experimentRepository).getOne(ID);
+        when(userRepository.getReferenceById(ID)).thenReturn(user);
+        when(experimentRepository.getReferenceById(ID)).thenReturn(experiment1);
+        assertTrue(participantService.isInvalidParticipant(ID, ID, SECRET, true));
+        verify(userRepository).getReferenceById(ID);
+        verify(experimentRepository).getReferenceById(ID);
         verify(participantRepository).findByUserAndExperiment(user, experiment1);
     }
 
     @Test
     public void testIsInvalidParticipantEntityNotFound() {
-        when(userRepository.getOne(ID)).thenReturn(user);
-        when(experimentRepository.getOne(ID)).thenReturn(experiment1);
+        when(userRepository.getReferenceById(ID)).thenReturn(user);
+        when(experimentRepository.getReferenceById(ID)).thenReturn(experiment1);
         when(participantRepository.findByUserAndExperiment(user, experiment1)).thenThrow(EntityNotFoundException.class);
-        assertTrue(participantService.isInvalidParticipant(ID, ID, SECRET));
-        verify(userRepository).getOne(ID);
-        verify(experimentRepository).getOne(ID);
+        assertTrue(participantService.isInvalidParticipant(ID, ID, SECRET, true));
+        verify(userRepository).getReferenceById(ID);
+        verify(experimentRepository).getReferenceById(ID);
         verify(participantRepository).findByUserAndExperiment(user, experiment1);
     }
 
     @Test
     public void testIsInvalidParticipantSecretBlank() {
         assertThrows(IllegalArgumentException.class,
-                () -> participantService.isInvalidParticipant(ID, ID, " ")
+                () -> participantService.isInvalidParticipant(ID, ID, " ", true)
         );
-        verify(userRepository, never()).getOne(anyInt());
-        verify(experimentRepository, never()).getOne(anyInt());
+        verify(userRepository, never()).getReferenceById(anyInt());
+        verify(experimentRepository, never()).getReferenceById(anyInt());
         verify(participantRepository, never()).findByUserAndExperiment(any(), any());
     }
 
     @Test
     public void testIsInvalidParticipantSecretNull() {
         assertThrows(IllegalArgumentException.class,
-                () -> participantService.isInvalidParticipant(ID, ID, null)
+                () -> participantService.isInvalidParticipant(ID, ID, null, true)
         );
-        verify(userRepository, never()).getOne(anyInt());
-        verify(experimentRepository, never()).getOne(anyInt());
+        verify(userRepository, never()).getReferenceById(anyInt());
+        verify(experimentRepository, never()).getReferenceById(anyInt());
         verify(participantRepository, never()).findByUserAndExperiment(any(), any());
     }
 
     @Test
     public void testIsInvalidParticipantInvalidExperimentId() {
         assertThrows(IllegalArgumentException.class,
-                () -> participantService.isInvalidParticipant(ID, 0, SECRET)
+                () -> participantService.isInvalidParticipant(ID, 0, SECRET, true)
         );
-        verify(userRepository, never()).getOne(anyInt());
-        verify(experimentRepository, never()).getOne(anyInt());
+        verify(userRepository, never()).getReferenceById(anyInt());
+        verify(experimentRepository, never()).getReferenceById(anyInt());
         verify(participantRepository, never()).findByUserAndExperiment(any(), any());
     }
 
     @Test
     public void testIsInvalidParticipantInvalidUserId() {
         assertThrows(IllegalArgumentException.class,
-                () -> participantService.isInvalidParticipant(-9, ID, SECRET)
+                () -> participantService.isInvalidParticipant(-9, ID, SECRET, true)
         );
-        verify(userRepository, never()).getOne(anyInt());
-        verify(experimentRepository, never()).getOne(anyInt());
+        verify(userRepository, never()).getReferenceById(anyInt());
+        verify(experimentRepository, never()).getReferenceById(anyInt());
         verify(participantRepository, never()).findByUserAndExperiment(any(), any());
     }
 
