@@ -19,14 +19,12 @@ import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.context.annotation.Import;
 import org.springframework.http.MediaType;
-import org.springframework.mock.web.MockHttpServletRequest;
-import org.springframework.security.web.csrf.CsrfToken;
-import org.springframework.security.web.csrf.HttpSessionCsrfTokenRepository;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.servlet.MockMvc;
 
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyBoolean;
 import static org.mockito.ArgumentMatchers.anyInt;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.never;
@@ -85,9 +83,6 @@ public class EventRestControllerIntegrationTest {
             return new byte[]{1, 2, 3};
         }
     };
-    private final String TOKEN_ATTR_NAME = "org.springframework.security.web.csrf.HttpSessionCsrfTokenRepository.CSRF_TOKEN";
-    private final HttpSessionCsrfTokenRepository httpSessionCsrfTokenRepository = new HttpSessionCsrfTokenRepository();
-    private final CsrfToken csrfToken = httpSessionCsrfTokenRepository.generateToken(new MockHttpServletRequest());
 
     @BeforeEach
     public void setup() throws JSONException {
@@ -168,26 +163,22 @@ public class EventRestControllerIntegrationTest {
     public void testStoreBlockEvent() throws Exception {
         mvc.perform(post("/store/block")
                         .content(blockEventObject.toString())
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.APPLICATION_JSON)
                         .accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk());
-        verify(participantService).isInvalidParticipant(USER_ID, Experiment_ID, SECRET);
+        verify(participantService).isInvalidParticipant(USER_ID, Experiment_ID, SECRET, true);
         verify(eventService).saveBlockEvent(any());
     }
 
     @Test
     public void testStoreBlockEventInvalidParticipant() throws Exception {
-        when(participantService.isInvalidParticipant(USER_ID, Experiment_ID, SECRET)).thenReturn(true);
+        when(participantService.isInvalidParticipant(USER_ID, Experiment_ID, SECRET, true)).thenReturn(true);
         mvc.perform(post("/store/block")
                         .content(blockEventObject.toString())
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.APPLICATION_JSON)
                         .accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk());
-        verify(participantService).isInvalidParticipant(USER_ID, Experiment_ID, SECRET);
+        verify(participantService).isInvalidParticipant(USER_ID, Experiment_ID, SECRET, true);
         verify(eventService, never()).saveBlockEvent(any());
     }
 
@@ -196,12 +187,10 @@ public class EventRestControllerIntegrationTest {
         blockEventObject.put("experiment", "no");
         mvc.perform(post("/store/block")
                         .content(blockEventObject.toString())
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.APPLICATION_JSON)
                         .accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk());
-        verify(participantService, never()).isInvalidParticipant(anyInt(), anyInt(), anyString());
+        verify(participantService, never()).isInvalidParticipant(anyInt(), anyInt(), anyString(), anyBoolean());
         verify(eventService, never()).saveBlockEvent(any());
     }
 
@@ -209,12 +198,10 @@ public class EventRestControllerIntegrationTest {
     public void testStoreClickEvent() throws Exception {
         mvc.perform(post("/store/click")
                         .content(clickEventObject.toString())
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.APPLICATION_JSON)
                         .accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk());
-        verify(participantService).isInvalidParticipant(USER_ID, Experiment_ID, SECRET);
+        verify(participantService).isInvalidParticipant(USER_ID, Experiment_ID, SECRET, true);
         verify(eventService).saveClickEvent(any());
     }
 
@@ -223,12 +210,10 @@ public class EventRestControllerIntegrationTest {
         clickEventObject.put("time", "-1");
         mvc.perform(post("/store/click")
                         .content(clickEventObject.toString())
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.APPLICATION_JSON)
                         .accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk());
-        verify(participantService, never()).isInvalidParticipant(anyInt(), anyInt(), anyString());
+        verify(participantService, never()).isInvalidParticipant(anyInt(), anyInt(), anyString(), anyBoolean());
         verify(eventService, never()).saveClickEvent(any());
     }
 
@@ -236,12 +221,10 @@ public class EventRestControllerIntegrationTest {
     public void testStoreDebuggerEvent() throws Exception {
         mvc.perform(post("/store/debugger")
                         .content(debuggerEventObject.toString())
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.APPLICATION_JSON)
                         .accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk());
-        verify(participantService).isInvalidParticipant(USER_ID, Experiment_ID, SECRET);
+        verify(participantService).isInvalidParticipant(USER_ID, Experiment_ID, SECRET, true);
         verify(eventService).saveDebuggerEvent(any());
     }
 
@@ -250,12 +233,10 @@ public class EventRestControllerIntegrationTest {
         debuggerEventObject.put("time", ".");
         mvc.perform(post("/store/debugger")
                         .content(debuggerEventObject.toString())
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.APPLICATION_JSON)
                         .accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk());
-        verify(participantService, never()).isInvalidParticipant(anyInt(), anyInt(), anyString());
+        verify(participantService, never()).isInvalidParticipant(anyInt(), anyInt(), anyString(), anyBoolean());
         verify(eventService, never()).saveDebuggerEvent(any());
     }
 
@@ -263,26 +244,22 @@ public class EventRestControllerIntegrationTest {
     public void testStoreQuestionEvent() throws Exception {
         mvc.perform(post("/store/question")
                         .content(questionEventObject.toString())
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.APPLICATION_JSON)
                         .accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk());
-        verify(participantService).isInvalidParticipant(USER_ID, Experiment_ID, SECRET);
+        verify(participantService).isInvalidParticipant(USER_ID, Experiment_ID, SECRET, true);
         verify(eventService).saveQuestionEvent(any());
     }
 
     @Test
     public void testStoreQuestionEventInvalidParticipant() throws Exception {
-        when(participantService.isInvalidParticipant(USER_ID, Experiment_ID, SECRET)).thenReturn(true);
+        when(participantService.isInvalidParticipant(USER_ID, Experiment_ID, SECRET, true)).thenReturn(true);
         mvc.perform(post("/store/question")
                         .content(questionEventObject.toString())
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.APPLICATION_JSON)
                         .accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk());
-        verify(participantService).isInvalidParticipant(USER_ID, Experiment_ID, SECRET);
+        verify(participantService).isInvalidParticipant(USER_ID, Experiment_ID, SECRET, true);
         verify(eventService, never()).saveQuestionEvent(any());
     }
 
@@ -290,12 +267,10 @@ public class EventRestControllerIntegrationTest {
     public void testStoreResourceEvent() throws Exception {
         mvc.perform(post("/store/resource")
                         .content(resourceEventObject.toString())
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.APPLICATION_JSON)
                         .accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk());
-        verify(participantService).isInvalidParticipant(USER_ID, Experiment_ID, SECRET);
+        verify(participantService).isInvalidParticipant(USER_ID, Experiment_ID, SECRET, true);
         verify(eventService).saveResourceEvent(any());
     }
 
@@ -304,12 +279,10 @@ public class EventRestControllerIntegrationTest {
         resourceEventObject.put("time", "user");
         mvc.perform(post("/store/resource")
                         .content(resourceEventObject.toString())
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.APPLICATION_JSON)
                         .accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk());
-        verify(participantService, never()).isInvalidParticipant(anyInt(), anyInt(), anyString());
+        verify(participantService, never()).isInvalidParticipant(anyInt(), anyInt(), anyString(), anyBoolean());
         verify(eventService, never()).saveResourceEvent(any());
     }
 
@@ -317,12 +290,10 @@ public class EventRestControllerIntegrationTest {
     public void testStoreFileEvent() throws Exception {
         mvc.perform(post("/store/file")
                         .content(fileEventObject.toString())
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.APPLICATION_JSON)
                         .accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk());
-        verify(participantService).isInvalidParticipant(USER_ID, Experiment_ID, SECRET);
+        verify(participantService).isInvalidParticipant(USER_ID, Experiment_ID, SECRET, true);
         verify(fileService).saveFile(any());
     }
 
@@ -331,12 +302,10 @@ public class EventRestControllerIntegrationTest {
         fileEventObject.put("file", "%");
         mvc.perform(post("/store/file")
                         .content(fileEventObject.toString())
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.APPLICATION_JSON)
                         .accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk());
-        verify(participantService, never()).isInvalidParticipant(anyInt(), anyInt(), anyString());
+        verify(participantService, never()).isInvalidParticipant(anyInt(), anyInt(), anyString(), anyBoolean());
         verify(fileService, never()).saveFile(any());
     }
 
@@ -344,26 +313,22 @@ public class EventRestControllerIntegrationTest {
     public void testStoreZipFile() throws Exception {
         mvc.perform(post("/store/zip")
                         .content(sb3ZipObject.toString())
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.APPLICATION_JSON)
                         .accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk());
-        verify(participantService).isInvalidParticipant(USER_ID, Experiment_ID, SECRET);
+        verify(participantService).isInvalidParticipant(USER_ID, Experiment_ID, SECRET, true);
         verify(fileService).saveSb3Zip(any());
     }
 
     @Test
     public void testStoreZipFileInvalidParticipant() throws Exception {
-        when(participantService.isInvalidParticipant(USER_ID, Experiment_ID, SECRET)).thenReturn(true);
+        when(participantService.isInvalidParticipant(USER_ID, Experiment_ID, SECRET, true)).thenReturn(true);
         mvc.perform(post("/store/zip")
                         .content(sb3ZipObject.toString())
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.APPLICATION_JSON)
                         .accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk());
-        verify(participantService).isInvalidParticipant(USER_ID, Experiment_ID, SECRET);
+        verify(participantService).isInvalidParticipant(USER_ID, Experiment_ID, SECRET, true);
         verify(fileService, never()).saveSb3Zip(any());
     }
 
@@ -372,12 +337,10 @@ public class EventRestControllerIntegrationTest {
         sb3ZipObject.put("user", "unicorn");
         mvc.perform(post("/store/zip")
                         .content(sb3ZipObject.toString())
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.APPLICATION_JSON)
                         .accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk());
-        verify(participantService, never()).isInvalidParticipant(anyInt(), anyInt(), anyString());
+        verify(participantService, never()).isInvalidParticipant(anyInt(), anyInt(), anyString(), anyBoolean());
         verify(fileService, never()).saveSb3Zip(any());
     }
 
@@ -386,12 +349,10 @@ public class EventRestControllerIntegrationTest {
         when(experimentService.getSb3File(Experiment_ID)).thenReturn(experimentProjection);
         mvc.perform(post("/store/sb3")
                         .content(dataObject.toString())
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.APPLICATION_JSON)
                         .accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk());
-        verify(participantService).isInvalidParticipant(USER_ID, Experiment_ID, SECRET);
+        verify(participantService).isInvalidParticipant(USER_ID, Experiment_ID, SECRET, true);
         verify(experimentService).getSb3File(Experiment_ID);
     }
 
@@ -415,12 +376,10 @@ public class EventRestControllerIntegrationTest {
         });
         mvc.perform(post("/store/sb3")
                         .content(dataObject.toString())
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.APPLICATION_JSON)
                         .accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isNotFound());
-        verify(participantService).isInvalidParticipant(USER_ID, Experiment_ID, SECRET);
+        verify(participantService).isInvalidParticipant(USER_ID, Experiment_ID, SECRET, true);
         verify(experimentService).getSb3File(Experiment_ID);
     }
 
@@ -429,12 +388,10 @@ public class EventRestControllerIntegrationTest {
         when(experimentService.getSb3File(Experiment_ID)).thenThrow(NotFoundException.class);
         mvc.perform(post("/store/sb3")
                         .content(dataObject.toString())
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.APPLICATION_JSON)
                         .accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isNotFound());
-        verify(participantService).isInvalidParticipant(USER_ID, Experiment_ID, SECRET);
+        verify(participantService).isInvalidParticipant(USER_ID, Experiment_ID, SECRET, true);
         verify(experimentService).getSb3File(Experiment_ID);
     }
 
@@ -443,12 +400,10 @@ public class EventRestControllerIntegrationTest {
         when(eventService.findFirstJSON(USER_ID, Experiment_ID)).thenReturn(JSON);
         mvc.perform(post("/store/json")
                         .content(dataObject.toString())
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.APPLICATION_JSON)
                         .accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk());
-        verify(participantService).isInvalidParticipant(USER_ID, Experiment_ID, SECRET);
+        verify(participantService).isInvalidParticipant(USER_ID, Experiment_ID, SECRET, true);
         verify(eventService).findFirstJSON(USER_ID, Experiment_ID);
     }
 
@@ -456,12 +411,10 @@ public class EventRestControllerIntegrationTest {
     public void testRetrieveLastJsonJsonNull() throws Exception {
         mvc.perform(post("/store/json")
                         .content(dataObject.toString())
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.APPLICATION_JSON)
                         .accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isNotFound());
-        verify(participantService).isInvalidParticipant(USER_ID, Experiment_ID, SECRET);
+        verify(participantService).isInvalidParticipant(USER_ID, Experiment_ID, SECRET, true);
         verify(eventService).findFirstJSON(USER_ID, Experiment_ID);
     }
 
@@ -470,26 +423,22 @@ public class EventRestControllerIntegrationTest {
         when(eventService.findFirstJSON(USER_ID, Experiment_ID)).thenThrow(NotFoundException.class);
         mvc.perform(post("/store/json")
                         .content(dataObject.toString())
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.APPLICATION_JSON)
                         .accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isNotFound());
-        verify(participantService).isInvalidParticipant(USER_ID, Experiment_ID, SECRET);
+        verify(participantService).isInvalidParticipant(USER_ID, Experiment_ID, SECRET, true);
         verify(eventService).findFirstJSON(USER_ID, Experiment_ID);
     }
 
     @Test
     public void testRetrieveLastJsonInvalidParticipant() throws Exception {
-        when(participantService.isInvalidParticipant(USER_ID, Experiment_ID, SECRET)).thenReturn(true);
+        when(participantService.isInvalidParticipant(USER_ID, Experiment_ID, SECRET, true)).thenReturn(true);
         mvc.perform(post("/store/json")
                         .content(dataObject.toString())
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.APPLICATION_JSON)
                         .accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isBadRequest());
-        verify(participantService).isInvalidParticipant(USER_ID, Experiment_ID, SECRET);
+        verify(participantService).isInvalidParticipant(USER_ID, Experiment_ID, SECRET, true);
         verify(eventService, never()).findFirstJSON(anyInt(), anyInt());
     }
 }
