@@ -27,10 +27,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.http.MediaType;
-import org.springframework.mock.web.MockHttpServletRequest;
 import org.springframework.security.test.context.support.WithMockUser;
-import org.springframework.security.web.csrf.CsrfToken;
-import org.springframework.security.web.csrf.HttpSessionCsrfTokenRepository;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.servlet.MockMvc;
@@ -41,6 +38,7 @@ import java.util.List;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.nullValue;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyBoolean;
 import static org.mockito.ArgumentMatchers.anyInt;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.never;
@@ -106,9 +104,6 @@ public class HomeControllerIntegrationTest {
             "password", "");
     private final Page<ExperimentTableProjection> experimentPage = new PageImpl<>(getExperimentProjections(5));
     private final Page<CourseTableProjection> coursePage = new PageImpl<>(getCourseTableProjections(3));
-    private final String TOKEN_ATTR_NAME = "org.springframework.security.web.csrf.HttpSessionCsrfTokenRepository.CSRF_TOKEN";
-    private final HttpSessionCsrfTokenRepository httpSessionCsrfTokenRepository = new HttpSessionCsrfTokenRepository();
-    private final CsrfToken csrfToken = httpSessionCsrfTokenRepository.generateToken(new MockHttpServletRequest());
 
     @BeforeEach
     public void setup() {
@@ -123,8 +118,6 @@ public class HomeControllerIntegrationTest {
     @Test
     public void testGetIndexPage() throws Exception {
         mvc.perform(get("/")
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.ALL)
                         .accept(MediaType.ALL))
                 .andExpect(status().isOk())
@@ -141,8 +134,6 @@ public class HomeControllerIntegrationTest {
         when(pageService.getExperimentPage(any(PageRequest.class))).thenReturn(experimentPage);
         when(pageService.getCoursePage(any(PageRequest.class))).thenReturn(coursePage);
         mvc.perform(get("/")
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.ALL)
                         .accept(MediaType.ALL))
                 .andExpect(status().isOk())
@@ -166,8 +157,6 @@ public class HomeControllerIntegrationTest {
                 anyInt())).thenReturn(experimentPage);
         when(pageService.getCourseParticipantPage(any(PageRequest.class), anyInt())).thenReturn(coursePage);
         mvc.perform(get("/")
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.ALL)
                         .accept(MediaType.ALL))
                 .andExpect(status().isOk())
@@ -188,8 +177,6 @@ public class HomeControllerIntegrationTest {
     public void testGetIndexPageParticipantNotFound() throws Exception {
         when(userService.getUser(userDTO.getUsername())).thenThrow(NotFoundException.class);
         mvc.perform(get("/")
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.ALL)
                         .accept(MediaType.ALL))
                 .andExpect(status().is3xxRedirection())
@@ -208,8 +195,6 @@ public class HomeControllerIntegrationTest {
         when(pageService.getCoursePage(any(PageRequest.class))).thenReturn(coursePage);
         mvc.perform(get(PAGE_COURSE)
                         .param(PAGE_PARAM, CURRENT)
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.ALL)
                         .accept(MediaType.ALL))
                 .andExpect(status().isOk())
@@ -231,8 +216,6 @@ public class HomeControllerIntegrationTest {
         when(pageService.getCourseParticipantPage(any(PageRequest.class), anyInt())).thenReturn(coursePage);
         mvc.perform(get(PAGE_COURSE)
                         .param(PAGE_PARAM, CURRENT)
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.ALL)
                         .accept(MediaType.ALL))
                 .andExpect(status().isOk())
@@ -253,8 +236,6 @@ public class HomeControllerIntegrationTest {
         when(pageService.getLastCoursePage(userDTO.getId())).thenReturn(lastPage);
         mvc.perform(get(PAGE_COURSE)
                         .param(PAGE_PARAM, INVALID_NUMBER)
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.ALL)
                         .accept(MediaType.ALL))
                 .andExpect(status().is4xxClientError())
@@ -272,8 +253,6 @@ public class HomeControllerIntegrationTest {
         when(pageService.getExperimentPage(any(PageRequest.class))).thenReturn(experimentPage);
         mvc.perform(get(PAGE_EXPERIMENT)
                         .param(PAGE_PARAM, CURRENT)
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.ALL)
                         .accept(MediaType.ALL))
                 .andExpect(status().isOk())
@@ -294,8 +273,6 @@ public class HomeControllerIntegrationTest {
         when(pageService.getExperimentParticipantPage(any(PageRequest.class), anyInt())).thenReturn(experimentPage);
         mvc.perform(get(PAGE_EXPERIMENT)
                         .param(PAGE_PARAM, CURRENT)
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.ALL)
                         .accept(MediaType.ALL))
                 .andExpect(status().isOk())
@@ -316,8 +293,6 @@ public class HomeControllerIntegrationTest {
         when(pageService.getLastExperimentPage(userDTO.getId())).thenReturn(pageNum);
         mvc.perform(get(PAGE_EXPERIMENT)
                         .param(PAGE_PARAM, CURRENT)
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.ALL)
                         .accept(MediaType.ALL))
                 .andExpect(status().is4xxClientError())
@@ -334,8 +309,6 @@ public class HomeControllerIntegrationTest {
         when(userService.getUser(userDTO.getUsername())).thenThrow(NotFoundException.class);
         mvc.perform(get(PAGE_EXPERIMENT)
                         .param(PAGE_PARAM, CURRENT)
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.ALL)
                         .accept(MediaType.ALL))
                 .andExpect(status().is4xxClientError())
@@ -352,8 +325,6 @@ public class HomeControllerIntegrationTest {
         when(pageService.computeLastExperimentPage()).thenReturn(lastPage);
         mvc.perform(get(PAGE_EXPERIMENT)
                         .param(PAGE_PARAM, PAGE_PARAM)
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.ALL)
                         .accept(MediaType.ALL))
                 .andExpect(status().is4xxClientError())
@@ -370,8 +341,6 @@ public class HomeControllerIntegrationTest {
                         .param(EXPERIMENT, ID_STRING)
                         .param(USER, ID_STRING)
                         .param(SECRET, SECRET)
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.ALL)
                         .accept(MediaType.ALL))
                 .andExpect(status().isOk())
@@ -379,7 +348,7 @@ public class HomeControllerIntegrationTest {
                 .andExpect(model().attribute(EXPERIMENT, is(ID)))
                 .andExpect(model().attribute(USER, is(ID)))
                 .andExpect(view().name(FINISH));
-        verify(participantService).isInvalidParticipant(ID, ID, SECRET);
+        verify(participantService).isInvalidParticipant(ID, ID, SECRET, false);
         verify(experimentService).getExperiment(ID);
     }
 
@@ -390,14 +359,12 @@ public class HomeControllerIntegrationTest {
                         .param(EXPERIMENT, ID_STRING)
                         .param(USER, ID_STRING)
                         .param(SECRET, SECRET)
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.ALL)
                         .accept(MediaType.ALL))
                 .andExpect(status().is3xxRedirection())
                 .andExpect(model().attribute(THANKS, nullValue()))
                 .andExpect(view().name(Constants.ERROR));
-        verify(participantService).isInvalidParticipant(ID, ID, SECRET);
+        verify(participantService).isInvalidParticipant(ID, ID, SECRET, false);
         verify(experimentService).getExperiment(ID);
     }
 
@@ -407,14 +374,12 @@ public class HomeControllerIntegrationTest {
                         .param(EXPERIMENT, INVALID_NUMBER)
                         .param(USER, ID_STRING)
                         .param(SECRET, SECRET)
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.ALL)
                         .accept(MediaType.ALL))
                 .andExpect(status().is3xxRedirection())
                 .andExpect(model().attribute(THANKS, nullValue()))
                 .andExpect(view().name(Constants.ERROR));
-        verify(participantService, never()).isInvalidParticipant(anyInt(), anyInt(), anyString());
+        verify(participantService, never()).isInvalidParticipant(anyInt(), anyInt(), anyString(), anyBoolean());
         verify(experimentService, never()).getExperiment(anyInt());
     }
 
@@ -424,14 +389,12 @@ public class HomeControllerIntegrationTest {
                         .param(EXPERIMENT, ID_STRING)
                         .param(USER, BLANK)
                         .param(SECRET, SECRET)
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.ALL)
                         .accept(MediaType.ALL))
                 .andExpect(status().is3xxRedirection())
                 .andExpect(model().attribute(THANKS, nullValue()))
                 .andExpect(view().name(Constants.ERROR));
-        verify(participantService, never()).isInvalidParticipant(anyInt(), anyInt(), anyString());
+        verify(participantService, never()).isInvalidParticipant(anyInt(), anyInt(), anyString(), anyBoolean());
         verify(experimentService, never()).getExperiment(anyInt());
     }
 
@@ -439,8 +402,6 @@ public class HomeControllerIntegrationTest {
     public void testGetResetPage() throws Exception {
         MailServerSetter.setMailServer(true);
         mvc.perform(get("/reset")
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.ALL)
                         .accept(MediaType.ALL))
                 .andExpect(status().isOk())
@@ -451,8 +412,6 @@ public class HomeControllerIntegrationTest {
     public void testGetResetPageNoMailServer() throws Exception {
         MailServerSetter.setMailServer(false);
         mvc.perform(get("/reset")
-                        .sessionAttr(TOKEN_ATTR_NAME, csrfToken)
-                        .param(csrfToken.getParameterName(), csrfToken.getToken())
                         .contentType(MediaType.ALL)
                         .accept(MediaType.ALL))
                 .andExpect(status().is3xxRedirection())
